# src/main/java/com/example/bankcards/dto/JwtAuthenticationResponse.java
package com.example.bankcards.dto;

import io.swagger.v3.oas.annotations.media.Schema;
import lombok.Getter;
import lombok.Setter;

/**
 * Ответ c токеном доступа.
 */
@Getter
@Setter
@Schema(description = "Ответ c токеном доступа")
public class JwtAuthenticationResponse {

    @Schema(description = "Токен доступа")
    private String token;
}